Feat kafka consumer add (정산로직)

+++ DataMigration/src/main/java/com/datamigration/datamigration/batch/domain/Aggregation.java
@@ -22,5 +22,5 @@ public class Aggregation {
     private Long communityId;
 
     @Column(name = "points")
-    private Integer points;
+    private Long points;
 }
